dataengg/etlairflowstyle/requirements.txt
pandas
apache-airflow
pymysql
requests
streamlit
pydantic-settings
sqlalchemy
folium
streamlit_folium